create utils class for image / alphabetic image loading
